+++ src/src_dotnet/JAStudio.Core/Note/Sentences/ParsedMatch.cs
@@ -5,16 +5,16 @@ namespace JAStudio.Core.Note.Sentences;
 
 public class ParsedMatch
 {
-    public const int MissingNoteId = -1;
+    public const long MissingNoteId = -1;
     public static readonly ParsedWordSerializer Serializer = new();
 
     public int StartIndex { get; set; }
     public bool IsDisplayed { get; set; }
     public string Variant { get; set; }
     public string ParsedForm { get; set; }
-    public int VocabId { get; set; }
+    public long VocabId { get; set; }
 
-    public ParsedMatch(string variant, int startIndex, bool isDisplayed, string word, int vocabId)
+    public ParsedMatch(string variant, int startIndex, bool isDisplayed, string word, long vocabId)
     {
         StartIndex = startIndex;
         IsDisplayed = isDisplayed;
